MAINT: reorganize tests to match module structure
